requirements.txt
imageio==2.9.0
numpy==1.19.4
Pillow==8.0.1
scikit-learn==0.23.2
scipy==1.5.4
torch==1.7.0
torchvision==0.8.1
tqdm==4.53.0
wandb==0.12.2
matplotlib
